Make all (?) fields & rings attribute storing (#1690)
While I don't have an immediate application, all other field and ring
types in Nemo (and most in Oscar) are attribute storing (be it via an
explicit @attributes, or because they are singleton types). So it seems
sensible to also enable it for these types proactively. Otherwise
we end up writing "generic" code using attributes that doesn't work on
all kinds of rings.